pyproject.toml
[project]
name = "dag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-duckdb",
    "dagster-docker",
    "dagster-aws",
    "pandas",
    "pykrx",
    "fastparquet",
    "psycopg2-binary",
    "libsql-experimental",
    "requests",
    "beautifulsoup4",
    "chardet",
    "psutil",
    "boto3",
    "Pillow",
    "dagster-openai",
    "openai",
    "dagster-gemini",
    "opencv-python-headless"
]

[project.optional-dependencies]
dev = [
    "dagster-webserver", 
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dag.definitions"
code_location_name = "dag"

[tool.setuptools.packages.find]
exclude=["dag_tests"]
